requirements.txt
requests>=2.32.3
tensorflow>=2.18.0
pillow>=11.1.0